refactor(validators): move processError validator to helpers directory
Consolidate validation error processing logic in a central location under infrastructure/helpers/validators. This improves maintainability by having all validator utilities in one place and makes the import paths more consistent across the codebase.
