+++ src/infrastructure/external/auth/validators/request-reset-password-email.validator.ts
@@ -4,7 +4,7 @@ import globalStrings from "@/config/strings/global.strings.json";
 import { EMAIL_BASIC_REGEX, URL_REGEX } from "@/config/regex/validations.regex";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
 import { TRawJson } from "@/domain/shared/interfaces/general.interfaces";
-import { processValidationError } from "../../../validators/utils/processError.validator";
+import { processValidationError } from "../../../helpers/validators/processError.validator";
 
 const { VALIDATION } = globalStrings.ERRORS.AUTH.REQUEST_RESET_PASSWORD_EMAIL;
