+++ src/infrastructure/helpers/validators/processError.validator.ts
